add DirectionTab.vue
